requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions-durable
cryptography==3.0
Pillow==8.2.0
PyPDF2==1.26.0
requests==2.25.1
fuzzywuzzy==0.18.0
azure-cosmosdb-table==1.0.6
azure-storage-blob==2.1.0
fuzzywuzzy==0.18.0
pandas==1.3.1
python-dateutil==2.8.2